requirements.txt
# Core
numpy==1.26.4
scipy==1.11.4
pandas==2.1.4

# ML
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2
faiss-cpu==1.7.4

# NLP
transformers==4.30.2
sentence-transformers==2.2.2
tokenizers==0.13.3
nltk==3.9.2

# API
fastapi==0.103.2
starlette==0.27.0
pydantic==2.4.2
uvicorn==0.23.2

# Utils
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.1
